fix: rename tsx components to index
