+++ src/components/BoardCard/BoardCardProps.ts
