+++ src/components/DeleteModal/DeleteModalProps.ts
