+++ src/components/index.tsx
@@ -1,0 +1,6 @@
+export { Header } from './Header';
+export { BoardList } from './BoardList';
+export { BoardCard } from './BoardCard';
+export { ChangeModal } from './ChangeModal';
+export { CreateModal } from './CreateModal';
+export { DeleteModal } from './DeleteModal';
