+++ src/elements/Buttons/Search/index.tsx
@@ -1,6 +1,6 @@
 import React, { FC, useState } from 'react';
 import { Search } from './Search.styles';
-import { SearchProps } from '../../../utils/Props/SearchProps';
+import { SearchProps } from './SearchProps';
 
 export const SearchInput:FC<SearchProps> = ({ onSearch }) => {
   const [searchText, setSearchText] = useState('');
