+++ src/elements/index.tsx
@@ -1,0 +1,1 @@
+export { SearchInput } from './Buttons/Search'
